requirements.txt
accelerate
ipykernel
ipywidgets
numpy
pandas
python-dotenv
slackweb
transformers
torch
